# apps/backend/src/main/java/com/rendaxx/labs/domain/Driver.java
package com.rendaxx.labs.domain;

import jakarta.annotation.Nullable;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import jakarta.persistence.UniqueConstraint;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor(access = AccessLevel.PRIVATE)
@Entity
@Table(
        name = "drivers",
        uniqueConstraints = {@UniqueConstraint(name = "uk_drivers_passport", columnNames = "passport")})
public class Driver {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @NotBlank
    private String firstName;

    @Nullable
    private String middleName;

    @Size(min = 1, message = "must contain at least one non-whitespace character")
    @Pattern(regexp = ".*\\S.*", message = "must not be blank when provided")
    private String lastName;

    @NotBlank
    private String passport;
}