# specs/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Information Systems Labs API
  version: 1.0.0
  description: |
    REST API for managing routes, vehicles, retail points, drivers, orders and related resources.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/orders:
    post:
      tags: [Orders]
      summary: Create an order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveOrder'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Orders]
      summary: List orders
      operationId: listOrders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/OrderSort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paged list of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders]
      summary: Get order by id
      operationId: getOrder
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Orders]
      summary: Update order
      operationId: updateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveOrder'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Orders]
      summary: Delete order
      operationId: deleteOrder
      responses:
        '204':
          description: Order removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/drivers:
    post:
      tags: [Drivers]
      summary: Create a driver
      operationId: createDriver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveDriver'
      responses:
        '201':
          description: Driver created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Drivers]
      summary: List drivers
      operationId: listDrivers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/DriverSort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paged list of drivers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDriver'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/drivers/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Drivers]
      summary: Get driver by id
      operationId: getDriver
      responses:
        '200':
          description: Driver found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Drivers]
      summary: Update driver
      operationId: updateDriver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveDriver'
      responses:
        '200':
          description: Driver updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Drivers]
      summary: Delete driver
      operationId: deleteDriver
      responses:
        '204':
          description: Driver removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/retail-points:
    post:
      tags: [Retail Points]
      summary: Create a retail point
      operationId: createRetailPoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRetailPoint'
      responses:
        '201':
          description: Retail point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetailPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Retail Points]
      summary: List retail points
      operationId: listRetailPoints
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/RetailPointSort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paged list of retail points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRetailPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/retail-points/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Retail Points]
      summary: Get retail point by id
      operationId: getRetailPoint
      responses:
        '200':
          description: Retail point found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetailPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Retail Points]
      summary: Update retail point
      operationId: updateRetailPoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRetailPoint'
      responses:
        '200':
          description: Retail point updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetailPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Retail Points]
      summary: Delete retail point
      operationId: deleteRetailPoint
      responses:
        '204':
          description: Retail point removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/retail-points/{id}/nearest:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/Limit'
    get:
      tags: [Retail Points]
      summary: Get nearest retail points
      operationId: getNearestRetailPoints
      responses:
        '200':
          description: List of nearest retail points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetailPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
...
    OrderSort:
      name: sort
      in: query
      description: >
        Повторяемый параметр. Поля: id, goodsType, minTemperature, maxTemperature,
        volumeInCubicMeters, weightInKg. Формат: "field[,asc|desc]".
      schema:
        type: array
        items:
          type: string
          pattern: '^(id|goodsType|minTemperature|maxTemperature|volumeInCubicMeters|weightInKg)(,(?i:asc|desc))?$'
      style: form
      explode: true
    DriverSort:
      name: sort
      in: query
      description: >
        Повторяемый параметр. Поля: id, firstName, lastName, middleName, passport.
      schema:
        type: array
        items:
          type: string
          pattern: '^(id|firstName|lastName|middleName|passport)(,(?i:asc|desc))?$'
      style: form
      explode: true
...
    RetailPointSort:
      name: sort
      in: query
      description: >
        Повторяемый параметр. Поля: id, name, address, timezone, type.
      schema:
        type: array
        items:
          type: string
          pattern: '^(id|name|address|timezone|type)(,(?i:asc|desc))?$'
      style: form
      explode: true
...
    Filter:
      name: filter
      in: query
      style: deepObject
      explode: true
      schema:
        type: object
        additionalProperties:
          type: string
    Limit:
      name: limit
      in: query
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema:
            type: string
  schemas:
    Driver:
      type: object
      required: [id, firstName, lastName, passport]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        firstName:
          type: string
          minLength: 1
        middleName:
          type: string
          nullable: true
        lastName:
          type: string
          minLength: 1
        passport:
          type: string
          minLength: 1
    SaveDriver:
      type: object
      required: [firstName, lastName, passport]
      properties:
        firstName:
          type: string
          minLength: 1
        middleName:
          type: string
          nullable: true
        lastName:
          type: string
          minLength: 1
        passport:
          type: string
          minLength: 1
    Order:
      type: object
      required: [id, goodsType, minTemperature, maxTemperature, volumeInCubicMeters, weightInKg]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        goodsType:
          type: string
          minLength: 1
        minTemperature:
          type: integer
          format: int32
          minimum: -2147483648
          maximum: 2147483647
        maxTemperature:
          type: integer
          format: int32
          minimum: -2147483648
          maximum: 2147483647
        volumeInCubicMeters:
          type: number
          format: double
          minimum: 0.001
        weightInKg:
          type: number
          format: double
          minimum: 0.001
    SaveOrder:
      type: object
      required: [goodsType, minTemperature, maxTemperature, volumeInCubicMeters, weightInKg]
      properties:
        goodsType:
          type: string
          minLength: 1
        minTemperature:
          type: integer
          format: int32
          minimum: -2147483648
          maximum: 2147483647
        maxTemperature:
          type: integer
          format: int32
          minimum: -2147483648
          maximum: 2147483647
        volumeInCubicMeters:
          type: number
          format: double
          minimum: 0.001
        weightInKg:
          type: number
          format: double
          minimum: 0.001
    Point:
      type: object
      required: [longitude, latitude]
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
    RetailPoint:
      type: object
      required: [id, name, address, location, type, timezone]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        location:
          $ref: '#/components/schemas/Point'
        type:
          $ref: '#/components/schemas/PointType'
        timezone:
          type: string
          minLength: 1
    SaveRetailPoint:
      type: object
      required: [name, address, location, type, timezone]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        location:
          $ref: '#/components/schemas/Point'
        type:
          $ref: '#/components/schemas/PointType'
        timezone:
          type: string
          minLength: 1
...
    PageOrder:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Order'
    PageDriver:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Driver'
...
    PageRetailPoint:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RetailPoint'
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages, sort, first, last, empty]
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    PointType:
      type: string
      enum: [SHOP, WAREHOUSE, GARAGE]